carrerabahn-main/requirements.txt
pyserial
pyrealsense2
opencv-python
numpy
torch
matplotlib

# Optional for temp.py

# gymnasium
